Fix coordinate order in all entries
Strangely enough, the map on GitHub (and on the GeoJSONLint site) used the
coordinates in the other order, namely latitude and longitude were swapped.
This put all of the entries somewhere in the sea in the middle east instead
of somewhere in northern Germany!  This commit should fix the issue in the
data; the hotzen_spotz editor still needs to be corrected.